title-demo: pin PathFollow constants + s32 waypoint width; path_node WIP
Advances the title-demo-rider port arc (docs/title_writer_chains.md
"Port target — minimum viable"). Static + dynamic RE this session:

- **Constants pinned** from static ROM pool at 0x003CF4F0..0x003CF514
  (via new Ghidra ReadWord.py):
    kMaxYawStep   = 267  (DAT_003CF4F4 = 0x0000010B)
    kPathSpeed    = 8.0f (DAT_003CF4F8 = 0x41000000) ✓ matches JIT
    arriveDist    = 8.0f (bit-compare vs 0x41000000 after vsqrt)
    snap speed_xz = 0.0f (DAT_003CF500)

- **Waypoint coord width — s32, NOT s16.** Disasm at 0x003CF3D4..
  0x003CF408 shows `ldr r0,[r2,#0x18]; vcvt.f32.s32 s2,s0`. Grezzo
  widened path-node coords from N64 Vec3s. z_actor.c PathFollow_Update
  corrected — earlier draft's (s16) reads were wrong; port must use
  (s32*)(node+K).

- **path_node pointer — partial**. Static: FindDataWriters on the sole
  .data slot (0x00526DE8) storing PathFollow_Update's addr shows 0 xrefs;
  ListCallers confirms dispatch is via fn-ptr load (no BL Ghidra can see).
  Dynamic: candidate 0x098f4010 (stored in actor+0x128 of the settled
  title-demo rider at 0x09906A80) shifted actor yaw by Δ=+3010/60f when
  poked (soh3d/scratch/confirm_pathnode_by_poke.py) — 375× natural
  noise. NOT verified as the real path_node (struct's +0x1C=0x03020FFF
  makes no sense as a game-y-coord). Two candidate resolutions in doc.

- New Ghidra scripts: Disasm.py, FnAt.py, ReadWord.py. Compose with
  ListCallers/FindDataWriters for full static-RE workflow.

Follow-on scope items (shot-1 arc close ≠ title-parity close):
  - probe scriptedDelta across shots (FUN_00461324 writer)
  - RE FUN_00418B88 shot dispatch (5 shots, 5 matrix producers)
  - extend title-parity harness for multi-shot sampling

+++ src/code/z_actor.c
@@ -136,7 +136,10 @@ void Actor_TurnToPoint(void* actor, const Vec3f* target, s32 max_step) {
  * the effective "title-demo spline driver".
  *
  * Structural sequence:
- *   1. Decode target Vec3f from s16 waypoint coords at path_node+0x18/1C/20.
+ *   1. Decode target Vec3f from **s32** waypoint coords at path_node+0x18/1C/20.
+ *      Verified via disasm 0x003cf3d4..0x003cf408: `ldr r0,[r2,#0x18]` +
+ *      `vmov s0,r0` + `vcvt.f32.s32 s2,s0`. NOT s16 as an earlier draft
+ *      assumed — Grezzo widened path-node coords from N64 Vec3s to s32.
  *   2. Compute distance from actor->world.pos to target.
  *   3. If dist < 8.0f  →  snap to target, speed = 0 (arrived).
  *      Else           →  Actor_TurnToPoint(actor, &target, kMaxYawStep),
@@ -157,9 +160,10 @@ void PathFollow_Update(void* actor, void* param_2, void* path_node) {
     u8* node = (u8*)path_node;
 
     Vec3f target;
-    target.x = (f32)(*(s16*)(node + 0x18));
-    target.y = (f32)(*(s16*)(node + 0x1C));
-    target.z = (f32)(*(s16*)(node + 0x20));
+    /* WHOLE s32 → f32, not just low s16 — see body comment above. */
+    target.x = (f32)(*(s32*)(node + 0x18));
+    target.y = (f32)(*(s32*)(node + 0x1C));
+    target.z = (f32)(*(s32*)(node + 0x20));
 
     f32 dx = *(f32*)(base + ACTOR_WORLD_POS_X_OFF) - target.x;
     f32 dy = *(f32*)(base + ACTOR_WORLD_POS_Y_OFF) - target.y;
